Add metrics for workload resources
Every time series in Prometheus is tagged with a "job" label[1]. In
Kubernetes land, a job corresponds to the workload controller, such as a
deployment. This adds queries for metrics for the typical workload
controllers and associates them with the corresponding Kubernetes
object.

This is useful for autoscaling metrics based on quantiles, which don't
make sense as a target for averageValue.

From `kubectl api-resources --api-group=apps`:

    NAME           SHORTNAMES   KIND
    daemonsets     ds           DaemonSet
    deployments    deploy       Deployment
    statefulsets   sts          StatefulSet

[1]: https://prometheus.io/docs/concepts/jobs_instances/